добавлена очистка ошибок формы логина
