+++ packages/client/src/common-styles/Form.module.scss
@@ -10,3 +10,13 @@
   text-align: center;
   padding-top: 20px;
 }
+
+.link {
+  color: var(--color-text-on-dark);
+  text-align: center;
+
+  &:hover {
+    color: var(--color-accent);
+    transition: 0.4s;
+  }
+}
