+++ packages/client/src/components/Title/Title.tsx
@@ -1,11 +1,9 @@
-import { FC } from "react";
-import styles from "./Title.module.scss"
+import { FC } from 'react'
+import styles from './Title.module.scss'
 
 interface Title {
   text: string
 }
 export const Title: FC<Title> = ({ text }) => {
-  return (
-    <h2 className={styles.title}>{text}</h2>
-  )
+  return <h2 className={styles.title}>{text}</h2>
 }
