--- packages/client/src/pages/Registration/Registration.module.scss
@@ -1,9 +1,0 @@
-.link {
-  color: var(--color-text-on-dark);
-  text-align: center;
-
-  &:hover {
-    color: var(--color-accent);
-    transition: 0.4s;
-  }
-}
